Added data_retrieve and save_data
data_retrieve returns the data requested as a single column data frame. Automatically refers to the correct file, so just put in any valid label.
save_data saves a dataframe as a .csv. The file for now is saved as just "Saven_m", where n and m are the index